+++ src/com/scs/testgame/entities/TestGameClientPlayersAvatar.java
@@ -22,7 +22,7 @@ public class TestGameClientPlayersAvatar extends ClientPlayersAvatar {
 	private static final float PLAYER_RAD = 0.2f;
 	private static final float WEIGHT = 3f;
 */
-	public TestGameClientPlayersAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, HUD _hud, int eid, float x, float y, float z, byte side) {
+	public TestGameClientPlayersAvatar(AbstractGameClient _module, int _playerID, IInputDevice _input, Camera _cam, HUD _hud, int eid, float x, float y, float z, int side) {
 		super(_module, _playerID, _input, _cam, _hud, eid, x, y, z, side);
 	}
